+++ example/server.js
@@ -6,7 +6,7 @@ var App         = require('./client');
 express()
   .get('/bundle.js', browserify(__dirname + '/client', {debug: true, watch: true}))
   .get('/', function(req, res, next) {
-    ReactAsync.renderComponentToString(App(), function(err, markup, data) {
+    ReactAsync.renderComponentToStringWithAsyncState(App(), function(err, markup, data) {
       if (err) return next(err);
 
       markup = ReactAsync.injectIntoMarkup(markup, data, ['./bundle.js']);