+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.6.0
+
+  - Fibers are now optional, they are only needed if you want to pre-render
+    React components by fetching async state recursively, e.g. using
+    ReactAsync.renderComponentToStringWithAsyncState
+
+  - Remove ReactAsync.createClass, React.createClass with ReactAsync.Mixin
+    should be used instead.
+
+  - Remove ReactAsync.renderComponent, React.renderComponent should be used
+    instead
+
+  - Rename ReactAsync.renderComponentToString to
+    ReactAsync.renderComponentToStringWithAsyncState
+
+  - Add ReactAsync.isAsyncComponent
+
+  - Add ReactAsync.prefetchAsyncState
+
 ## 0.5.1
 
   - Check if async component is still mounted before updating its state from and